+++ types/affinda.d.ts
@@ -968,7 +968,7 @@ export declare interface AffindaAPIGetAllIndexesOptionalParams extends coreClien
     /** The numbers of results to return. */
     limit?: number;
     /** Filter indices by a document type */
-    documentType?: Enum16;
+    documentType?: Enum17;
 }
 
 /** Contains response data for the getAllIndexes operation. */
@@ -1343,6 +1343,8 @@ export declare interface AffindaAPIUpdateOrganizationOptionalParams extends core
     /** Used to sign webhook payloads so you can verify their integrity. */
     resthookSignatureKey?: string;
     name?: string;
+    /** Configuration of the embeddable validation tool. */
+    validationToolConfig?: ValidationToolConfig;
 }
 
 /** Contains response data for the updateOrganization operation. */
@@ -2067,24 +2069,24 @@ export declare interface EducationSearchScoreComponent {
 }
 
 /**
- * Defines values for Enum16. \
- * {@link KnownEnum16} can be used interchangeably with Enum16,
+ * Defines values for Enum17. \
+ * {@link KnownEnum17} can be used interchangeably with Enum17,
  *  this enum contains the known values that the service supports.
  * ### Known values supported by the service
  * **resumes** \
  * **job_descriptions**
  */
-export declare type Enum16 = string;
+export declare type Enum17 = string;
 
 /**
- * Defines values for Enum19. \
- * {@link KnownEnum19} can be used interchangeably with Enum19,
+ * Defines values for Enum20. \
+ * {@link KnownEnum20} can be used interchangeably with Enum20,
  *  this enum contains the known values that the service supports.
  * ### Known values supported by the service
  * **resumes** \
  * **job_descriptions**
  */
-export declare type Enum19 = string;
+export declare type Enum20 = string;
 
 export declare type ExpectedRemunerationAnnotation = Annotation & {
     parsed?: ExpectedRemunerationAnnotationParsed;
@@ -2683,14 +2685,14 @@ export declare enum KnownDocumentState {
     Rejected = "rejected"
 }
 
-/** Known values of {@link Enum16} that the service accepts. */
-export declare enum KnownEnum16 {
+/** Known values of {@link Enum17} that the service accepts. */
+export declare enum KnownEnum17 {
     Resumes = "resumes",
     JobDescriptions = "job_descriptions"
 }
 
-/** Known values of {@link Enum19} that the service accepts. */
-export declare enum KnownEnum19 {
+/** Known values of {@link Enum20} that the service accepts. */
+export declare enum KnownEnum20 {
     Resumes = "resumes",
     JobDescriptions = "job_descriptions"
 }
@@ -2798,6 +2800,12 @@ export declare enum KnownResumeSkillSourcesItemSection {
     SkillsInterestsLanguages = "Skills/Interests/Languages",
     TrainingCertifications = "Training/Certifications",
     ExtracurricularsLeadership = "Extracurriculars/Leadership"
+}
+
+/** Known values of {@link ThemeConfigPaletteMode} that the service accepts. */
+export declare enum KnownThemeConfigPaletteMode {
+    Light = "light",
+    Dark = "dark"
 }
 
 /** Known values of {@link Version} that the service accepts. */
@@ -2948,6 +2956,8 @@ export declare interface Organization {
     /** Used to sign webhook payloads so you can verify their integrity. */
     resthookSignatureKey?: string;
     isTrial?: boolean;
+    /** Configuration of the embeddable validation tool. */
+    validationToolConfig?: OrganizationValidationToolConfig;
 }
 
 export declare interface OrganizationCreate {
@@ -2987,6 +2997,8 @@ export declare interface OrganizationUpdate {
     avatar?: coreRestPipeline.RequestBodyType;
     /** Used to sign webhook payloads so you can verify their integrity. */
     resthookSignatureKey?: string;
+    /** Configuration of the embeddable validation tool. */
+    validationToolConfig?: ValidationToolConfig;
 }
 
 /**
@@ -2998,6 +3010,13 @@ export declare interface OrganizationUpdate {
  * **member**
  */
 export declare type OrganizationUserRole = string;
+
+/** Configuration of the embeddable validation tool. */
+export declare interface OrganizationValidationToolConfig {
+    theme?: ThemeConfig;
+    /** Hide the confirm document button and other actions. */
+    hideActions?: boolean;
+}
 
 export declare interface PageMeta {
     id: number;
@@ -3022,6 +3041,13 @@ export declare interface PaginatedResponse {
     previous?: string;
 }
 
+export declare interface PaletteColorOptions {
+    main: string;
+    light?: string;
+    dark?: string;
+    contrastText?: string;
+}
+
 export declare type Paths18Wh2VcV3InvitationsGetResponses200ContentApplicationJsonSchema = PaginatedResponse & PathsKhpbbuV3InvitationsGetResponses200ContentApplicationJsonSchemaAllof1 & {};
 
 export declare interface Paths1Czpnk1V3ResumeSearchEmbedPostRequestbodyContentApplicationJsonSchema {
@@ -3044,7 +3070,7 @@ export declare interface Paths1Qojy9V3ResthookSubscriptionsGetResponses200Conten
 
 export declare interface Paths1TvfqeiV3IndexPostResponses201ContentApplicationJsonSchema {
     name?: string;
-    documentType?: Enum19;
+    documentType?: Enum20;
 }
 
 export declare interface Paths2Ld2HiV3WorkspaceMembershipsGetResponses200ContentApplicationJsonSchemaAllof1 {
@@ -3759,6 +3785,56 @@ export declare type TextAnnotation = Annotation & {
     parsed?: string;
 };
 
+export declare interface ThemeConfig {
+    palette?: ThemeConfigPalette;
+    typography?: ThemeConfigTypography;
+    borderRadius?: number;
+    fontUrl?: string;
+}
+
+export declare interface ThemeConfigPalette {
+    mode?: ThemeConfigPaletteMode;
+    background?: ThemeConfigPaletteBackground;
+    text?: ThemeConfigPaletteText;
+    divider?: string;
+    primary?: PaletteColorOptions;
+    secondary?: PaletteColorOptions;
+    success?: PaletteColorOptions;
+    annotation?: PaletteColorOptions;
+    error?: PaletteColorOptions;
+    info?: PaletteColorOptions;
+    warning?: PaletteColorOptions;
+}
+
+export declare interface ThemeConfigPaletteBackground {
+    default?: string;
+    paper?: string;
+}
+
+/**
+ * Defines values for ThemeConfigPaletteMode. \
+ * {@link KnownThemeConfigPaletteMode} can be used interchangeably with ThemeConfigPaletteMode,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **light** \
+ * **dark**
+ */
+export declare type ThemeConfigPaletteMode = string;
+
+export declare interface ThemeConfigPaletteText {
+    primary?: string;
+    secondary?: string;
+    disabled?: string;
+}
+
+export declare interface ThemeConfigTypography {
+    fontFamily?: string;
+    fontSize?: string;
+    fontWeightRegular?: string;
+    fontWeightMedium?: string;
+    fontWeightBold?: string;
+}
+
 export declare interface User {
     /** Uniquely identify a user. */
     id?: number;
@@ -3777,6 +3853,13 @@ export declare interface UserNullable {
     email?: string;
     /** URL of the user's avatar. */
     avatar?: string;
+}
+
+/** Configuration of the embeddable validation tool. */
+export declare interface ValidationToolConfig {
+    theme?: ThemeConfig;
+    /** Hide the confirm document button and other actions. */
+    hideActions?: boolean;
 }
 
 /**