+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add `Organization.validationToolConfig` for configuration of the embeddable validation tool
 
 ## [6.5.2] - 2023-04-20
 ### Fixed